video-pipeline/requirements.txt
opencv-python==4.8.1.78
ffmpeg-python==0.2.0
numpy==1.24.3
Pillow==10.0.0
watchdog==3.0.0
python-dotenv==1.0.0
